fix(ipld): Fix critical error handling and split into submodules
This commit addresses critical error handling issues identified in PR review
and improves code organization by splitting the monolithic ipld.rs file.

**Error Handling Fixes:**
- CID conversion functions now return Result instead of panicking
- Optional field parsing returns errors for wrong types (not silent None)
- Integer casts include bounds checking with clear error messages
- Schema CID generation propagates relative_id parse errors

**Module Split:**
- ipld.rs (1047 lines) → ipld/ directory with focused submodules:
  - cid_convert.rs (43 lines) - CID version conversion
  - to_ipld.rs (248 lines) - Block/Delta → Ipld conversions
  - from_ipld.rs (589 lines) - Ipld → Block/Delta parsing
  - traversal.rs (100 lines) - IpldVisitor and link traversal
  - tests.rs (594 lines) - Comprehensive test coverage

**New Tests:**
- 7 FieldDefinition/CollectionDefinition roundtrip tests
- 11 error path tests for malformed IPLD input
- 2 helper method tests (is_definition, doc_id for definitions)

All 254+ tests pass, clippy clean, properly formatted.
